requirements.txt
asgiref==3.3.4
certifi==2021.10.8
cffi==1.14.5
charset-normalizer==2.0.9
cloudinary==1.28.0
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.7
whitenoise==5.3.0
